+++ src/z3c/dependencychecker/dependencychecker.py
@@ -11,6 +11,8 @@
 # FOR A PARTICULAR PURPOSE.
 #
 ##############################################################################
+import _ast
+import ast
 import commands
 import fnmatch
 import logging
@@ -24,6 +26,13 @@ import pkg_resources
 from z3c.dependencychecker import importchecker
 
 logger = logging.getLogger(__name__)
+
+PACKAGE_NAME_PATTERN = re.compile(r"""
+^            # Start of string
+[\w\-\.]+    # \w is a-z, A-Z, underscore, numbers.
+             # Dash is also ok, as is a dot.
+$            # End of string
+""", re.VERBOSE)
 
 ZCML_PACKAGE_PATTERN = re.compile(r"""
 \s           # Whitespace.
@@ -319,6 +328,49 @@ def includes_from_generic_setup_metadata(path):
     return modules, test_modules
 
 
+def includes_from_django_settings(path):
+    modules = []
+    test_modules = []
+    for path, dirs, files in os.walk(path):
+        for settingsfile in [os.path.abspath(os.path.join(path, filename))
+                             for filename in files
+                             if fnmatch.fnmatch(filename, '*settings.py')]:
+            contents = open(settingsfile).read()
+            found = []
+            parsed = ast.parse(open(settingsfile).read())
+            # We're looking for assignments like ``INSTALLED_APPS = ``.
+            assignments = [obj for obj in parsed.body
+                           if isinstance(obj, _ast.Assign)]
+            # We're looking for assignments with lists/tuples like
+            # ``INSTALLED_APPS = [a, b, c]``.
+            lists_or_tuples = [assignment.value for assignment in assignments
+                               if isinstance(assignment.value, _ast.List)
+                               or isinstance(assignment.value, _ast.Tuple)]
+            for list_or_tuple in lists_or_tuples:
+                strings = [getattr(element, 's', None)
+                           for element in list_or_tuple.elts]
+                if None in strings:
+                    # A tuple of languages or so: that has no 's' attribute.
+                    continue
+                suspect_strings = [s for s in strings
+                                   if not re.match(PACKAGE_NAME_PATTERN, s)]
+                if suspect_strings:
+                    # Something doesn't look like a package name.
+                    continue
+                found += strings
+            logger.debug(
+                "Found possible packages in Django-like settings file %s:",
+                settingsfile)
+            logger.debug(found)
+            # import pdb;pdb.set_trace()
+            if 'test' in settingsfile:
+                # testsettings.py, for instance.
+                test_modules += found
+            else:
+                modules += found
+    return modules, test_modules
+
+
 def imports_from_doctests(path):
     test_modules = []
     for path, dirs, files in os.walk(path):
@@ -403,20 +455,27 @@ def main():
     zcml_imports = db.resolvePkgNames(zcml_imports)
     zcml_test_imports = db.resolvePkgNames(zcml_test_imports)
 
+    (django_settings_imports,
+     django_settings_test_imports) = includes_from_django_settings(path)
+    django_settings_imports = db.resolvePkgNames(django_settings_imports)
+    django_settings_test_imports = db.resolvePkgNames(
+        django_settings_test_imports)
+
     doctest_imports = imports_from_doctests(path)
     (generic_setup_required, generic_setup_test_required) = \
         includes_from_generic_setup_metadata(path)
 
     print_unused_imports(unused_imports)
 
-    install_missing = filter_missing(install_imports + zcml_imports +
-                                     generic_setup_required,
-                                     install_required + stdlib)
+    install_missing = filter_missing(
+        install_imports + zcml_imports + generic_setup_required +
+        django_settings_imports,
+        install_required + stdlib)
     print_modules(install_missing, "Missing requirements")
 
     test_missing = filter_missing(
         test_imports + zcml_test_imports + doctest_imports +
-        generic_setup_test_required,
+        generic_setup_test_required + django_settings_test_imports,
         install_required + test_required + stdlib)
     print_modules(test_missing, "Missing test requirements")
 
@@ -442,6 +501,6 @@ def main():
 
     if install_missing or test_missing or install_unneeded or test_unneeded:
         print "Note: requirements are taken from the egginfo dir, so you need"
-        print "to re-run buildout (or setup.py or whatever) for changes in "
+        print "to re-run buildout (or setup.py or whatever) for changes in"
         print "setup.py to have effect."
         print